+++ ta4j-core/src/test/java/org/ta4j/core/IndicatorTest.java
@@ -1,7 +1,7 @@
 /*******************************************************************************
  *   The MIT License (MIT)
  *
- *   Copyright (c) 2014-2017 Marc de Verdelhan, 2017-2018 Ta4j Organization 
+ *   Copyright (c) 2014-2017 Marc de Verdelhan, 2017-2018 Ta4j Organization
  *   & respective authors (see AUTHORS)
  *
  *   Permission is hereby granted, free of charge, to any person obtaining a copy of
@@ -21,35 +21,42 @@
  *   IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  *   CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  *******************************************************************************/
-package org.ta4j.core.indicators.volume;
+package org.ta4j.core;
 
-import org.ta4j.core.Indicator;
-import org.ta4j.core.indicators.CachedIndicator;
-import org.ta4j.core.indicators.SMAIndicator;
+import java.util.ArrayList;
+import java.util.List;
+
+import org.junit.Test;
+import org.ta4j.core.mocks.MockIndicator;
 import org.ta4j.core.num.Num;
 
+import static org.junit.Assert.assertTrue;
+
 /**
- * The Moving volume weighted average price (MVWAP) Indicator.
- * @see <a href="http://www.investopedia.com/articles/trading/11/trading-with-vwap-mvwap.asp">
- *     http://www.investopedia.com/articles/trading/11/trading-with-vwap-mvwap.asp</a>
+ * Tests for {@link Indicator}.
+ *
+ * @author Johnny Lim
  */
-public class MVWAPIndicator extends CachedIndicator<Num> {
+public class IndicatorTest {
 
-    private final Indicator<Num> sma;
-    
-    /**
-     * Constructor.
-     * @param vwap the vwap
-     * @param barCount the time frame
-     */
-    public MVWAPIndicator(VWAPIndicator vwap, int barCount) {
-        super(vwap);
-        sma = new SMAIndicator(vwap, barCount);
-    }
+	@Test
+	public void toDouble() {
+		TimeSeries series = new BaseTimeSeries();
+		List<Num> values = new ArrayList<>();
+		for (int i = 0; i < 100; i++) {
+			values.add(series.numOf(i));
+		}
+		Indicator<Num> indicator = new MockIndicator(series, values);
 
-    @Override
-    protected Num calculate(int index) {
-        return sma.getValue(index);
-    }
+		int index = 10;
+		int barCount = 3;
+
+		Double[] doubles = Indicator.toDouble(indicator, index, barCount);
+
+		assertTrue(doubles.length == barCount);
+		assertTrue(doubles[0] == 8d);
+		assertTrue(doubles[1] == 9d);
+		assertTrue(doubles[2] == 10d);
+	}
 
 }