+++ ta4j-core/src/test/java/org/ta4j/core/analysis/criteria/TotalProfitCriterionTest.java
@@ -1,25 +1,26 @@
-/*
-  The MIT License (MIT)
-
-  Copyright (c) 2014-2017 Marc de Verdelhan, Ta4j Organization & respective authors (see AUTHORS)
-
-  Permission is hereby granted, free of charge, to any person obtaining a copy of
-  this software and associated documentation files (the "Software"), to deal in
-  the Software without restriction, including without limitation the rights to
-  use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
-  the Software, and to permit persons to whom the Software is furnished to do so,
-  subject to the following conditions:
-
-  The above copyright notice and this permission notice shall be included in all
-  copies or substantial portions of the Software.
-
-  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
-  FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
-  COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
-  IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
-  CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
+/*******************************************************************************
+ *   The MIT License (MIT)
+ *
+ *   Copyright (c) 2014-2017 Marc de Verdelhan, 2017-2018 Ta4j Organization 
+ *   & respective authors (see AUTHORS)
+ *
+ *   Permission is hereby granted, free of charge, to any person obtaining a copy of
+ *   this software and associated documentation files (the "Software"), to deal in
+ *   the Software without restriction, including without limitation the rights to
+ *   use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+ *   the Software, and to permit persons to whom the Software is furnished to do so,
+ *   subject to the following conditions:
+ *
+ *   The above copyright notice and this permission notice shall be included in all
+ *   copies or substantial portions of the Software.
+ *
+ *   THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ *   IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ *   FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ *   COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+ *   IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ *   CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ *******************************************************************************/
 package org.ta4j.core.analysis.criteria;
 
 import org.junit.Test;
@@ -30,6 +31,7 @@ import org.ta4j.core.num.Num;
 import java.util.function.Function;
 
 import static org.junit.Assert.*;
+import static org.ta4j.core.TestUtils.assertNumEquals;
 
 public class TotalProfitCriterionTest extends AbstractCriterionTest{
 
@@ -45,7 +47,7 @@ public class TotalProfitCriterionTest extends AbstractCriterionTest{
                 Order.buyAt(3,series), Order.sellAt(5,series));
 
         AnalysisCriterion profit = new TotalProfitCriterion();
-        assertEquals(1.10 * 1.05, profit.calculate(series, tradingRecord), TestUtils.GENERAL_OFFSET);
+        assertNumEquals(1.10 * 1.05, profit.calculate(series, tradingRecord));
     }
 
     @Test
@@ -56,7 +58,7 @@ public class TotalProfitCriterionTest extends AbstractCriterionTest{
                 Order.buyAt(2,series), Order.sellAt(5,series));
 
         AnalysisCriterion profit = new TotalProfitCriterion();
-        assertEquals(0.95 * 0.7, profit.calculate(series, tradingRecord), TestUtils.GENERAL_OFFSET);
+        assertNumEquals(0.95 * 0.7, profit.calculate(series, tradingRecord));
     }
 
     @Test
@@ -67,7 +69,7 @@ public class TotalProfitCriterionTest extends AbstractCriterionTest{
                 Order.sellAt(2,series), Order.buyAt(5,series));
 
         AnalysisCriterion profit = new TotalProfitCriterion();
-        assertEquals((1 / 0.95) * (1 / 0.7), profit.calculate(series, tradingRecord), TestUtils.GENERAL_OFFSET);
+        assertNumEquals((1 / 0.95) * (1 / 0.7), profit.calculate(series, tradingRecord));
     }
 
     @Test
@@ -75,7 +77,7 @@ public class TotalProfitCriterionTest extends AbstractCriterionTest{
         MockTimeSeries series = new MockTimeSeries(numFunction, 100, 95, 100, 80, 85, 70);
 
         AnalysisCriterion profit = new TotalProfitCriterion();
-        assertEquals(1d, profit.calculate(series, new BaseTradingRecord()), TestUtils.GENERAL_OFFSET);
+        assertNumEquals(1d, profit.calculate(series, new BaseTradingRecord()));
     }
 
     @Test
@@ -83,15 +85,15 @@ public class TotalProfitCriterionTest extends AbstractCriterionTest{
         MockTimeSeries series = new MockTimeSeries(numFunction, 100, 95, 100, 80, 85, 70);
         AnalysisCriterion profit = new TotalProfitCriterion();
         Trade trade = new Trade();
-        assertEquals(1d, profit.calculate(series, trade), TestUtils.GENERAL_OFFSET);
+        assertNumEquals(1d, profit.calculate(series, trade));
         trade.operate(0);
-        assertEquals(1d, profit.calculate(series, trade), TestUtils.GENERAL_OFFSET);
+        assertNumEquals(1d, profit.calculate(series, trade));
     }
 
     @Test
     public void betterThan() {
         AnalysisCriterion criterion = new TotalProfitCriterion();
-        assertTrue(criterion.betterThan(2.0, 1.5));
-        assertFalse(criterion.betterThan(1.5, 2.0));
+        assertTrue(criterion.betterThan(numOf(2.0), numOf(1.5)));
+        assertFalse(criterion.betterThan(numOf(1.5), numOf(2.0)));
     }
 }