+++ CHANGELOG.md
@@ -9,8 +9,11 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 
 ### Fixed
 - **TradingBotOnMovingTimeSeries**: fixed calculations and ArithmeticException Overflow
+- **Fixed wrong indexing in**: `Indicator.toDouble()`.
+
 ### Changed
 - **ALL INDICATORS**: `Decimal` replaced by `Num`.
+- **ALL CRITERION**: Calculations modified to use `Num`.
 - **AbstractIndicator**: new `AbstractIndicator#numOf(Number n)` function as counterpart of dropped `Decimal.valueOf(double|int|..)`
 - **TimeSeries | Bar**: preferred way to add bar data to a `TimeSeries` is directly to the series via new `TimeSeries#addBar(time,open,high,..)` functions. It ensures to use the correct `Num` implementation of the series
 - **XlsTestsUtils**: now processes xls with one or more days between data rows (daily, weekly, monthly, etc).  Also handle xls #DIV/0! calculated cells (imported as NaN.NaN)